提交代码

+++ txlcn-tm/src/main/java/com/codingapi/txlcn/tm/support/service/impl/TxExceptionServiceImpl.java
@@ -89,7 +89,7 @@ public class TxExceptionServiceImpl implements TxExceptionService {
     @Override
     public int transactionState(String groupId) {
         log.debug("transactionState > groupId: {}", groupId);
-        List<Integer> lists = txExceptionRepository.getTransactionStateByGroupId(groupId,PageRequest.of(0, 1));
+        List<Byte> lists = txExceptionRepository.getTransactionStateByGroupId(groupId,PageRequest.of(0, 1));
         if (lists.size()<=0) {
             return -1;
         }
